% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regicor.R
\name{calcular_regicor}
\alias{calcular_regicor}
\title{Càlcul de la puntuació regicor}
\usage{
calcular_regicor(df, noms_cols)
}
\arguments{
\item{df}{un dataframe amb els valors}

\item{noms_cols}{vector amb els noms de les columnes que s'utilitzaran pel càlcul. Vegeu detalls.}
}
\value{
Retorna un vector amb els valors de risc de regicor
}
\description{
Funció vectoritzada per calcular els valors de l'escala de risc regicor
}
\details{
El paràmetre noms_cols és un vector amb els noms de les columnes del teu dataframe que s'utilitzaran
per fer els càlculs. S'han d'afegir els noms de les columnes per aquest ordre: edat, sexe, fumar, colesterol total,
tensió sistòlica, tensió diastòlica, colesterol HDL i diabetis.
}
\keyword{cardiovascular,}
\keyword{regicor}
\keyword{risc}
